services/common/pyproject.toml
[project]
name = "chessbeast-common"
version = "0.1.0"
description = "Common utilities for ChessBeast Python services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/common"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4", "SIM"]
ignore = ["D100", "D104", "D107"]

[tool.mypy]
python_version = "3.12"
strict = true
